+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Exclude deleted posts when pulling posts
 - Exclude hidden attachment group when setup
 - Support batch pulling posts
+- Add publish parameters to Front-matter
 
 ## 1.0.0